requirements.txt
aioamqp==0.10.0
appdirs==1.4.3
msgpack-python==0.4.8
packaging==16.8
pyparsing==2.2.0
six==1.10.0
